pyproject.toml
[project]
name = "surf"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "httpx",
    "mcp[cli]",
    "python-dotenv",
]

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["surf"]
